refactor(server): fold command menu label sync into WorkflowCommonWorkspaceService
Move the workflow command menu label sync logic out of a dedicated service
and into WorkflowCommonWorkspaceService (which already owns command menu item
cleanup), drop the explanatory comment, and merge the dedicated integration
spec into the existing workflow-resolver integration suite.
