+++ packages/twenty-server/src/modules/workflow/common/workspace-services/workflow-common.workspace-service.ts
@@ -1,7 +1,9 @@
 import { Injectable, Logger } from '@nestjs/common';
 
 import { isDefined, isValidUuid } from 'twenty-shared/utils';
+import { In } from 'typeorm';
 
+import { type WorkspaceAuthContext } from 'src/engine/core-modules/auth/types/workspace-auth-context.type';
 import { CommandMenuItemService } from 'src/engine/metadata-modules/command-menu-item/command-menu-item.service';
 import { type FlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/types/flat-entity-maps.type';
 import { findFlatEntityByIdInFlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-id-in-flat-entity-maps.util';
@@ -36,6 +38,7 @@ import {
   WorkflowTriggerException,
   WorkflowTriggerExceptionCode,
 } from 'src/modules/workflow/workflow-trigger/exceptions/workflow-trigger.exception';
+import { getWorkflowCommandMenuItemLabel } from 'src/modules/workflow/workflow-trigger/utils/get-workflow-command-menu-item-label.util';
 
 export type ObjectMetadataInfo = {
   flatObjectMetadata: FlatObjectMetadata;
@@ -102,6 +105,77 @@ export class WorkflowCommonWorkspaceService {
     }
 
     return { ...workflowVersion, trigger: workflowVersion.trigger };
+  }
+
+  async syncCommandMenuItemLabelForWorkflows(
+    workflowIds: string[],
+    authContext: WorkspaceAuthContext,
+  ): Promise<void> {
+    const workspaceId = authContext.workspace?.id;
+
+    if (!isDefined(workspaceId) || workflowIds.length === 0) {
+      return;
+    }
+
+    const workflows =
+      await this.globalWorkspaceOrmManager.executeInWorkspaceContext(
+        async () => {
+          const workflowRepository =
+            await this.globalWorkspaceOrmManager.getRepository<WorkflowWorkspaceEntity>(
+              workspaceId,
+              'workflow',
+              { shouldBypassPermissionChecks: true },
+            );
+
+          return workflowRepository.find({
+            where: { id: In(workflowIds) },
+          });
+        },
+        authContext,
+      );
+
+    await Promise.all(
+      workflows.map((workflow) =>
+        this.syncCommandMenuItemLabelForWorkflow(workflow, workspaceId),
+      ),
+    );
+  }
+
+  private async syncCommandMenuItemLabelForWorkflow(
+    workflow: WorkflowWorkspaceEntity,
+    workspaceId: string,
+  ): Promise<void> {
+    if (!isDefined(workflow.lastPublishedVersionId)) {
+      return;
+    }
+
+    const existingCommandMenuItem =
+      await this.commandMenuItemService.findByWorkflowVersionId(
+        workflow.lastPublishedVersionId,
+        workspaceId,
+      );
+
+    if (!isDefined(existingCommandMenuItem)) {
+      return;
+    }
+
+    const label = getWorkflowCommandMenuItemLabel(workflow);
+
+    if (
+      existingCommandMenuItem.label === label &&
+      existingCommandMenuItem.shortLabel === label
+    ) {
+      return;
+    }
+
+    await this.commandMenuItemService.update(
+      {
+        id: existingCommandMenuItem.id,
+        label,
+        shortLabel: label,
+      },
+      workspaceId,
+    );
   }
 
   async getFlatEntityMaps(workspaceId: string): Promise<{
