+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/twenty-2.code-workspace
@@ -1,0 +1,20 @@
+{
+	"folders": [
+		{
+			"path": "../../../../../../.."
+		},
+		{
+			"path": "../../../../../../../../my-app"
+		}
+	],
+	"settings": {
+		"workbench.colorTheme": "Kimbie Dark",
+		"javascript.format.enable": false,
+		"typescript.format.enable": false,
+		"files.associations": {
+			".cursorrules": "markdown"
+		},
+		"typescript.tsdk": "node_modules/typescript/lib",
+		"typescript.experimental.useTsgo": false
+	}
+}
